refactor : refactor resolvers
