+++ src/resolvers/repository.ts
@@ -1,9 +1,13 @@
-import { Repository } from '../models/types';
-import { repoList } from '../__db__/mock';
+import { Repository } from '../models/Repository';
+import { get, tableNames } from "../__db__";
+
+const REPOSITORY: tableNames = 'Repository';
 
 const repo = {
-  repo: (params: any, args: { id: string }, context: any, info: any) =>
-    repoList.filter((repo: Repository) => repo.id === args.id)[0],
+  repo: async (params: any, args: { id: string }) => {
+    const result = await get({ table: REPOSITORY });
+    return result.filter((repo: Repository) => repo.id === args.id)[0];
+  }
 };
 
 export default repo;
